+++ src/autocomplete/dropdown.js
@@ -31,8 +31,12 @@ function Dropdown(o) {
   this.isOpen = false;
   this.isEmpty = true;
   this.minLength = o.minLength || 0;
-  this.cssClasses = _.mixin({}, css.defaultClasses, o.cssClasses || {});
   this.templates = {};
+  this.appendTo = o.appendTo || false;
+  this.css = _.mixin({}, css, o.appendTo ? css.appendTo : {});
+  this.cssClasses = o.cssClasses = _.mixin({}, css.defaultClasses, o.cssClasses || {});
+  this.cssClasses.prefix =
+    o.cssClasses.formattedPrefix || _.formatPrefix(this.cssClasses.prefix, this.cssClasses.noPrefix);
 
   // bound functions
   onSuggestionClick = _.bind(this._onSuggestionClick, this);
@@ -44,6 +48,8 @@ function Dropdown(o) {
     .on('click.aa', cssClass, onSuggestionClick)
     .on('mouseenter.aa', cssClass, onSuggestionMouseEnter)
     .on('mouseleave.aa', cssClass, onSuggestionMouseLeave);
+
+  this.$container = o.appendTo ? o.wrapper : this.$menu;
 
   if (o.templates && o.templates.header) {
     this.templates.header = _.templatify(o.templates.header);
@@ -73,6 +79,11 @@ function Dropdown(o) {
     this.templates.footer = _.templatify(o.templates.footer);
     this.$menu.append(this.templates.footer());
   }
+
+  var self = this;
+  DOM.element(window).resize(function() {
+    self._redraw();
+  });
 }
 
 // instance methods
@@ -162,15 +173,23 @@ _.mixin(Dropdown.prototype, EventEmitter, {
   },
 
   _hide: function() {
-    this.$menu.hide();
+    this.$container.hide();
   },
 
   _show: function() {
     // can't use jQuery#show because $menu is a span element we want
     // display: block; not dislay: inline;
-    this.$menu.css('display', 'block');
+    this.$container.css('display', 'block');
+
+    this._redraw();
 
     this.trigger('shown');
+  },
+
+  _redraw: function redraw() {
+    if (!this.isOpen || !this.appendTo) return;
+
+    this.trigger('redrawn');
   },
 
   _getSuggestions: function getSuggestions() {
@@ -182,12 +201,16 @@ _.mixin(Dropdown.prototype, EventEmitter, {
   },
 
   _setCursor: function setCursor($el, updateInput) {
-    $el.first().addClass(_.className(this.cssClasses.prefix, this.cssClasses.cursor, true));
+    $el.first()
+      .addClass(_.className(this.cssClasses.prefix, this.cssClasses.cursor, true))
+      .attr('aria-selected', 'true');
     this.trigger('cursorMoved', updateInput);
   },
 
   _removeCursor: function removeCursor() {
-    this._getCursor().removeClass(_.className(this.cssClasses.prefix, this.cssClasses.cursor, true));
+    this._getCursor()
+      .removeClass(_.className(this.cssClasses.prefix, this.cssClasses.cursor, true))
+      .removeAttr('aria-selected');
   },
 
   _moveCursor: function moveCursor(increment) {
@@ -272,7 +295,7 @@ _.mixin(Dropdown.prototype, EventEmitter, {
   },
 
   setLanguageDirection: function setLanguageDirection(dir) {
-    this.$menu.css(dir === 'ltr' ? css.ltr : css.rtl);
+    this.$menu.css(dir === 'ltr' ? this.css.ltr : this.css.rtl);
   },
 
   moveCursorUp: function moveCursorUp() {
@@ -295,6 +318,10 @@ _.mixin(Dropdown.prototype, EventEmitter, {
     }
 
     return datum;
+  },
+
+  getCurrentCursor: function getCurrentCursor() {
+    return this._getCursor().first();
   },
 
   getDatumForCursor: function getDatumForCursor() {