+++ src/common/utils.js
@@ -99,8 +99,12 @@ module.exports = {
 
   noop: function() {},
 
+  formatPrefix: function(prefix, noPrefix) {
+    return noPrefix ? '' : prefix + '-';
+  },
+
   className: function(prefix, clazz, skipDot) {
-    return (skipDot ? '' : '.') + prefix + '-' + clazz;
+    return (skipDot ? '' : '.') + prefix + clazz;
   },
 
   escapeHTML: function(str, originalHighlightTags) {