+++ test/unit/typeahead_spec.js
@@ -40,6 +40,41 @@ describe('Typeahead', function() {
     this.dropdown = this.view.dropdown;
   });
 
+  describe('appendTo', function() {
+    it('should throw if used with hint', function(done) {
+      expect(function() {
+        return new Typeahead({
+          input: this.$input,
+          hint: true,
+          appendTo: 'body'
+        });
+      }).toThrow();
+      done();
+    });
+
+    it('should be appended to the target of appendTo', function(done) {
+      var node = document.createElement('div');
+      document.querySelector('body').appendChild(node);
+
+      expect(node.children.length).toEqual(0);
+
+      this.view.destroy();
+
+      this.view = new Typeahead({
+        input: this.$input,
+        hint: false,
+        appendTo: node
+      });
+
+      expect(document.querySelectorAll('.algolia-autocomplete').length).toEqual(1);
+      expect(node.children.length).toEqual(1);
+
+      this.view.destroy();
+
+      done();
+    });
+  });
+
   describe('when dropdown triggers suggestionClicked', function() {
     beforeEach(function() {
       this.dropdown.getDatumForSuggestion.and.returnValue(testDatum);
@@ -85,6 +120,7 @@ describe('Typeahead', function() {
   describe('when dropdown triggers cursorMoved', function() {
     beforeEach(function() {
       this.dropdown.getDatumForCursor.and.returnValue(testDatum);
+      this.dropdown.getCurrentCursor.and.returnValue($('<div id="option-id"></div>'));
     });
 
     it('should update the input value', function() {
@@ -92,6 +128,13 @@ describe('Typeahead', function() {
 
       expect(this.input.setInputValue)
         .toHaveBeenCalledWith(testDatum.value, true);
+    });
+
+    it('should update the active descendant', function() {
+      this.dropdown.trigger('cursorMoved', false);
+
+      expect(this.input.setActiveDescendant)
+        .toHaveBeenCalledWith('option-id');
     });
 
     it('should not update the input', function() {
@@ -195,6 +238,21 @@ describe('Typeahead', function() {
 
       expect(spy).toHaveBeenCalled();
     });
+
+    it('should trigger autocomplete:redrawn', function() {
+      var spy;
+
+      this.$input.on('autocomplete:redrawn', spy = jasmine.createSpy());
+
+      this.dropdown.trigger('redrawn');
+
+      expect(spy).toHaveBeenCalled();
+    });
+
+    it('should set the input\'s aria-expanded to true', function() {
+      this.dropdown.trigger('opened');
+      expect(this.input.expand).toHaveBeenCalled();
+    });
   });
 
   describe('when dropdown triggers closed', function() {
@@ -212,6 +270,11 @@ describe('Typeahead', function() {
       this.dropdown.trigger('closed');
 
       expect(spy).toHaveBeenCalled();
+    });
+
+    it('should set the input\'s aria-expanded to false', function() {
+      this.dropdown.trigger('closed');
+      expect(this.input.collapse).toHaveBeenCalled();
     });
   });